Added more extensive usage of the config json file

+++ showboat/__init__.py
@@ -2,6 +2,7 @@ from util import *
 from pkg_resources import resource_filename
 import shutil
 
+
 def deposit_skeleton(dst_path):
     if not os.path.exists(dst_path):
         os.mkdir(dst_path)
@@ -14,7 +15,7 @@ def deposit_skeleton(dst_path):
     os.mkdir(os.path.join(dst_path, 'styles'))
 
 def compile(src_path, out_path, copy_assets=False, overwrite_output=True,
-            build_thumbnails=True, config=default_config()):
+            build_thumbnails=True):
     
     out_path = os.path.join(out_path, 'output')
     
@@ -28,30 +29,35 @@ def compile(src_path, out_path, copy_assets=False, overwrite_output=True,
     os.mkdir(out_path)
     
     
-    styles_path = resource_filename(__name__, 'payload/styles')
-    scripts_path = resource_filename(__name__, 'payload/scripts')
+    global_styles_path = resource_filename(__name__, 'payload/styles')
+    global_scripts_path = resource_filename(__name__, 'payload/scripts')
+    local_styles_path = os.path.join(src_path, 'styles')
+    local_scripts_path = os.path.join(src_path, 'scripts')
     
     # rsync in the supporting file directories
-    rsync(styles_path, out_path)
-    rsync(scripts_path, out_path)
+    rsync(global_styles_path, out_path)
+    rsync(global_scripts_path, out_path)
+    rsync(local_styles_path, out_path)
+    rsync(local_scripts_path, out_path)
     
     # compile the coffeescript
     coffeescript_compile(out_path)
     
     assets_path = os.path.abspath(config['assets_path'])
     
-    # symlink or copy in the assets
-    for item in os.listdir(assets_path):
-        if item == '.DS_Store':
-            continue
+    # symlink or copy in the global assets, if they exist
+    if os.path.exists(assets_path):
+        for item in os.listdir(assets_path):
+            if item == '.DS_Store':
+                continue
         
-        asset_src = os.path.join(assets_path, item)
-        asset_dst = os.path.join(out_path, item)
+            asset_src = os.path.join(assets_path, item)
+            asset_dst = os.path.join(out_path, item)
         
-        if copy_assets:
-            shutil.copytree(asset_src, asset_dst)
-        else:
-            os.symlink(asset_src, asset_dst)
+            if copy_assets:
+                shutil.copytree(asset_src, asset_dst)
+            else:
+                os.symlink(asset_src, asset_dst)
     
     # compile the jade
     jade_compile(src_path, out_path)
@@ -61,5 +67,5 @@ def compile(src_path, out_path, copy_assets=False, overwrite_output=True,
         # count the number of slides
         n_slides = count_n_slides(out_path)
         build_slide_thumbnails(out_path, n_slides)
-    
 
+
